moved serial communication and threshold storage to separate files
